technology/serializers.py
from .models import Technology
from rest_framework import serializers

class TechnologySerializer(serializers.ModelSerializer):
    class Meta:
        model = Technology
        fields = ['id', 'name']